% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{abstreams}
\alias{abstreams}
\title{Dataset: A-B Streams}
\format{A river network object, see \link{rivernetwork}}
\usage{
data(abstreams)
}
\description{
A complex river network object, a subset of the streams in the Absaroka-Beartooth Wilderness.
}
\keyword{datasets}

