% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{abstreams0}
\alias{abstreams0}
\title{Dataset: A-B Streams 0}
\format{A river network object, see \link{rivernetwork}}
\usage{
data(abstreams0)
}
\description{
An unusably messy river network object, included for the purpose of testing river network editing functions.
}
\keyword{datasets}
