% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverdist_1.R
\name{addcumuldist}
\alias{addcumuldist}
\title{Add Cumulative Distance to a River Network}
\usage{
addcumuldist(rivers)
}
\arguments{
\item{rivers}{The river network object to use.}
}
\value{
Returns an object of class \code{"rivernetwork"} containing all
  spatial and topological information.  See \link{rivernetwork-class}.
}
\description{
Adds a vector of cumulative distances to a river network.  Called internally.
}
\examples{
Gulk1 <- addcumuldist(rivers=Gulk)
}
\author{
Matt Tyers
}
