% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectroute_riverdistance.R
\name{detectterminal}
\alias{detectterminal}
\title{Detect terminal segments}
\usage{
detectterminal(rivers, map = FALSE)
}
\arguments{
\item{rivers}{The river network object to use}

\item{map}{Whether or not to draw a sanity-check map, showing the calculated 
route in entirety.  Defaults to FALSE.}
}
\description{
Detect terminal segments
}
\examples{
\dontrun{
data(Gulk)
plot(Gulk)
Gulk1 <- detectterminal(Gulk, map = TRUE)
Gulk1$terminal
}
}
