% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissolve_zoom.R
\name{dissolve}
\alias{dissolve}
\title{Dissolve}
\usage{
dissolve(rivers)
}
\arguments{
\item{rivers}{The river network object to use}
}
\value{
A new river network object with segments combined
}
\description{
Acts like a spatial dissolve within a GIS environment.  Simplifies a river network object by combining "runs" of segments with no other connections.
}
\note{
This function is called within \link{cleanup}, which is recommended in most cases.
}
\examples{
data(Kenai2)
plot(x=Kenai2)

Kenai2dissolve <- dissolve(rivers=Kenai2)
plot(x=Kenai2dissolve)
}
\author{
Matt Tyers
}
\seealso{
line2network
}

