% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{fakefish}
\alias{fakefish}
\title{Dataset: Fakefish}
\format{A data frame}
\usage{
data(fakefish)
}
\description{
A set of observations of Fakefish on the Gulkana River and its tributaries.
}
\details{
\itemize{ 
\item \code{x}. X-coordinate of observation (Alaska Albers Equal Area). Note that the locations do not align with the river network object. 
\item \code{y}. Y-coordinate of observation 
\item \code{seg}. River segment (with x- and y-coordinates snapped to river network object) 
\item \code{vert}. River vertex
\item \code{fish.id}. Numeric identifier for each fish (individual fish were observed more than once) 
\item \code{flight}. Numeric identifier for each telemetry flight 
\item \code{flight.date}. Date of each telemetry flight 
}
}
\seealso{
\link{Gulk}
}
\keyword{datasets}
