% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{fakefish_density}
\alias{fakefish_density}
\title{Dataset: Fakefish Density}
\format{
A river density object, see \link{riverdensity}, \link{plotriverdensity}, \link{riverdensity-class}
}
\usage{
data(fakefish_density)
}
\description{
An object created by \link{riverdensity}, describing the density of Fakefish points in the Gulkana River during ten surveys.
}
\details{
Intended for plotting using \link{plotriverdensity}.
}
\keyword{datasets}
