% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{Gulk}
\alias{Gulk}
\title{Dataset: Gulkana River}
\format{
A river network object, see \link{rivernetwork}
}
\usage{
data(Gulk)
}
\description{
A stretch of Gulkana River and tributaries.
}
\keyword{datasets}
