% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{Kenai1}
\alias{Kenai1}
\title{Dataset: Kenai River 1}
\format{A river network object, see \link{rivernetwork}}
\usage{
data(Kenai1)
}
\description{
A first pass at a messy river network object.
}
\seealso{
\link{Kenai2}, \link{Kenai3}
}
\keyword{datasets}

