% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{Kenai2}
\alias{Kenai2}
\title{Dataset: Kenai River 2}
\format{A river network object, see \link{rivernetwork}}
\usage{
data(Kenai2)
}
\description{
A second pass at a messy river network object.  In this iteration of cleanup,
several non-connected segments have been removed.
}
\seealso{
\link{Kenai1}, \link{Kenai3}
}
\keyword{datasets}

