% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{Kenai3}
\alias{Kenai3}
\title{Dataset: Kenai River 3}
\format{A river network object, see \link{rivernetwork}}
\usage{
data(Kenai3)
}
\description{
A third pass at a messy river network object.  In this iteration of cleanup,
several non-connected segments have been removed, and several series of
segments have been dissolved into single segments.
}
\seealso{
\link{Kenai1}, \link{Kenai2}
}
\keyword{datasets}

