% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{KilleyW}
\alias{KilleyW}
\title{Dataset: Killey River, West Channel}
\format{A river network object, see \link{rivernetwork}}
\usage{
data(KilleyW)
}
\description{
A messy and braided section of the Kenai River network - actually a subset of \link{Kenai3}.
}
\keyword{datasets}
