% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{Koyukuk0}
\alias{Koyukuk0}
\title{Dataset: Koyukuk River 0}
\format{A river network object, see \link{rivernetwork}}
\usage{
data(Koyukuk0)
}
\description{
An unusably messy river network object, included for the purpose of testing river network editing functions.
}
\seealso{
\link{Koyukuk1}, \link{Koyukuk2}
}
\keyword{datasets}
