% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{Koyukuk1}
\alias{Koyukuk1}
\title{Dataset: Koyukuk River 1}
\format{A river network object, see \link{rivernetwork}}
\usage{
data(Koyukuk1)
}
\description{
A first pass at a messy river network object.  The way it was dissolved in
ArcGIS makes the endpoints appear disconnected to \link{line2network} and the
topologies do not work.
}
\seealso{
\link{Koyukuk2}
}
\keyword{datasets}

