% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{Koyukuk2}
\alias{Koyukuk2}
\title{Dataset: Koyukuk River 2}
\format{A river network object, see \link{rivernetwork}}
\usage{
data(Koyukuk2)
}
\description{
A second pass at a messy river network object, with topologies fixed from \link{Koyukuk1}.
}
\seealso{
\link{Koyukuk1}
}
\keyword{datasets}
