% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{line98}
\alias{line98}
\title{Dataset: Line 98 of Kenai River 1 (Long-Lat)}
\format{
A matrix of values
}
\usage{
data(line98)
}
\description{
A matrix of coordinates in longitude-latitude, used to illustrate coordinate
transformation.  Coordinates come from arbitrary line number 98 in the Kenai
River 1 shapefile, rendered in long-lat.
}
\keyword{datasets}
