% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverdist_1.R
\name{pdist}
\alias{pdist}
\title{Pythagorean Distance}
\usage{
pdist(p1, p2)
}
\arguments{
\item{p1}{X-Y coordinates of point 1}

\item{p2}{X-Y coordinates of point 2}
}
\value{
Distance (numeric)
}
\description{
Pythagorean distance between two points.  Called internally.
}
\examples{
point1 <- c(1,3)
point2 <- c(4,7)

pdist(point1,point2)
}
\author{
Matt Tyers
}

