% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverdist_1.R
\name{pdisttot}
\alias{pdisttot}
\title{Total Pythagorean Distance}
\usage{
pdisttot(xy)
}
\arguments{
\item{xy}{A matrix of X-Y coordinates of the sequence of points.}
}
\value{
Distance (numeric)
}
\description{
Total Pythagorean distance of a sequence of points.  Called internally.
}
\examples{
points <- matrix(c(1:10), nrow=5, ncol=2, byrow=FALSE)

pdisttot(xy=points)
}
\author{
Matt Tyers
}
