% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apps_of_direction.R
\name{plotmatobslist}
\alias{plotmatobslist}
\title{Plot Upstream Distance Between Observations of All Individuals}
\usage{
plotmatobslist(matobslist, type = "boxplot", showN = TRUE, ...)
}
\arguments{
\item{matobslist}{A list of distance matrices returned from \link{matobslist}.}

\item{type}{If \code{type} is set to \code{"boxplot"}, boxplots will be
produced for each cell.  If \code{type} is set to \code{"confint"}, lines
denoting an approximate 95 percent confidence interval for the mean will be
produced instead.  If \code{type} is set to \code{"dotplot"}, a jittered
dotplot will be produced for each cell, which will be the most appropriate
if sample sizes are small.  Defaults to \code{"boxplot"}.}

\item{showN}{Whether to display the sample size for each cell.  Defaults to
TRUE.}

\item{...}{Additional plotting arguments.}
}
\description{
Produces a matrix of plots (boxplots are default), with plot \code{[i,j]} giving the
  distribution of upstream distances from observation \code{i} to observation
  \code{j}, for all individuals.
}
\note{
Building routes from the river mouth to each river network segment may
  greatly reduce computation time (see \link{buildsegroutes}).
}
\examples{
data(Gulk, smallset)
matobslist <- matobslist(unique=smallset$id, survey=smallset$flight, seg=smallset$seg,
   vert=smallset$vert, rivers=Gulk)
plotmatobslist(matobslist)
plotmatobslist(matobslist,type="confint")
plotmatobslist(matobslist,type="dotplot")

data(fakefish)
# matobslist <- matobslist(unique=fakefish$fish.id, survey=fakefish$flight, seg=fakefish$seg,
#   vert=fakefish$vert, rivers=Gulk)
# plotmatobslist(matobslist)
}
\author{
Matt Tyers
}
\seealso{
\link{upstream}, \link{upstreammatobs}
}

