% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apps_of_direction.R
\name{plotupstreammatobs}
\alias{plotupstreammatobs}
\title{Plot Upstream Distance Between Observations of All Individuals}
\usage{
plotupstreammatobs(ID, survey, seg, vert, rivers, type = "boxplot", ...)
}
\arguments{
\item{ID}{A vector of unique identifiers for each fish.}

\item{survey}{A vector of identifiers for each survey.  It is recommended to
use a numeric or date format (see \link{as.Date}) to preserve survey order.}

\item{seg}{A vector of river locations (segment component).}

\item{vert}{A vector of river locations (vertex component).}

\item{rivers}{The river network object to use.}

\item{type}{If \code{type} is set to \code{"boxplot"}, boxplots will be
produced for each cell.  If \code{type} is set to \code{"confint"}, lines
denoting an approximate 95% confidence interval for the mean will be
produced instead.  Defaults to \code{"boxplot"}.}

\item{...}{Additional arguments for \link{upstream}.}
}
\description{
Produces a matrix of boxplots, with plot \code{[i,j]} giving the
  distribution of upstream distances from observation \code{i} to observation
  \code{j}, for all individuals.  Each distance is calculated in
  \link{upstream}.
}
\note{
Building routes from the river mouth to each river network segment may
  greatly reduce computation time (see \link{buildsegroutes}).
}
\examples{
data(Gulk, smallset)
plotupstreammatobs(ID=smallset$id, survey=smallset$flight, seg=smallset$seg,
   vert=smallset$vert, rivers=Gulk)

data(fakefish)
# plotupstreammatobs(ID=fakefish$fish.id, survey=fakefish$flight, seg=fakefish$seg,
#   vert=fakefish$vert, rivers=Gulk)
}
\author{
Matt Tyers
}
\seealso{
\link{upstream}, \link{upstreammatobs}
}

