% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestroute_cleanup.R
\name{removeduplicates}
\alias{removeduplicates}
\title{Remove Duplicates}
\usage{
removeduplicates(rivers)
}
\arguments{
\item{rivers}{The river network object to use}
}
\value{
A new river network object with duplicated segments removed, see \link{rivernetwork}
}
\description{
Removes duplicated line segments, which can sometimes exist within a shapefile.
}
\examples{
data(abstreams0)
zoomtoseg(seg=c(170,171,157),rivers=abstreams0)

abstreams1 <- removeduplicates(rivers=abstreams0)
zoomtoseg(seg=c(166,167,154),rivers=abstreams1)
}
\author{
Matt Tyers
}
\seealso{
line2network
}

