% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/riverdist_1.R
\name{removeunconnected}
\alias{removeunconnected}
\title{Remove Unconnected Segments}
\usage{
removeunconnected(rivers)
}
\arguments{
\item{rivers}{The river network object to use.}
}
\description{
Detects and removes segments that are not connected to the river mouth.
}
\note{
This function is called within \link{cleanup}, which is recommended in most cases.
}
\examples{
data(Kenai1)
Kenai1.1 <- dissolve(Kenai1)
Kenai1.1 <- setmouth(seg=63,vert=40,rivers=Kenai1.1)
plot(Kenai1.1)

Kenai1.2 <- removeunconnected(Kenai1.1)
plot(Kenai1.2)
}
\author{
Matt Tyers
}

