% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_help.R
\docType{data}
\name{smallset}
\alias{smallset}
\title{Dataset: Smallset}
\format{
A data frame
}
\usage{
data(smallset)
}
\description{
A small set of observations of fakefish on the Gulkana River and its tributaries.
}
\details{
\itemize{ 
\item \code{x}. X-coordinate of observation (Alaska Albers Equal Area). Note that the locations do not align with the river network object. 
\item \code{y}. Y-coordinate of observation 
\item \code{seg}. River segment 
\item \code{vert}. River vertex
\item \code{fish.id}. Numeric identifier for each fish (individual fish were observed more than once) 
\item \code{flight}. Numeric identifier for each telemetry flight
}
}
\seealso{
\link{Gulk}
}
\keyword{datasets}
