% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segsplit.R
\name{splitsegmentat}
\alias{splitsegmentat}
\title{Split a Segment at a Specified Vertex}
\usage{
splitsegmentat(seg, vert, rivers)
}
\arguments{
\item{seg}{The segment to split}

\item{vert}{The vertex to split it at}

\item{rivers}{The river network object to use}
}
\value{
A new, updated river network object
}
\description{
Splits a segment at a specified vertex, creating two new segments.
}
\examples{
data(Gulk)
plot(x=Gulk)

Gulk2 <- splitsegmentat(seg=1, vert=400, rivers=Gulk)
plot(x=Gulk2)
}
\seealso{
\link{line2network}
}
\author{
Matt Tyers
}
