% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/riverdist_1.R
\name{topologydots}
\alias{topologydots}
\title{Check Connectivity of a River Network Object}
\usage{
topologydots(rivers, add = FALSE, ...)
}
\arguments{
\item{rivers}{The river network object to check}

\item{add}{Whether call a new plot (\code{FALSE}) or add dots to an existing
plot (\code{TRUE}).  Defaults to \code{FALSE}.}

\item{...}{Additional plotting arguments (see \link[graphics]{par})}
}
\description{
Produces a graphical check of the connectivity of a river
  network object.  It produces a \link{plot} of the river network object,
  and overlays red dots at non-connected endpoints and green dots at
  connected endpoints.
}
\examples{
data(Gulk)
topologydots(rivers=Gulk)
}
\author{
Matt Tyers
}
\seealso{
\link{line2network}
}

