% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverdist_1.R
\name{whoconnected}
\alias{whoconnected}
\title{Check Which Segments are Connected to a Given Segment.}
\usage{
whoconnected(seg, rivers)
}
\arguments{
\item{seg}{The segment to check}

\item{rivers}{The river network object it belongs to}
}
\value{
A vector of segment numbers
}
\description{
Returns which segments are connected to a specified segment within a river network.  It may be useful for error checking.
}
\examples{
data(Gulk)
plot(Gulk)
whoconnected(seg=4, rivers=Gulk)
}
\author{
Matt Tyers
}

