% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve_zoom.R
\name{zoomtoseg}
\alias{zoomtoseg}
\title{Zoom to segment}
\usage{
zoomtoseg(seg, rivers, ...)
}
\arguments{
\item{seg}{A segment or vector of segments to zoom to}

\item{rivers}{The river network object to use}

\item{...}{Additional plotting arguments (see \link[graphics]{par})}
}
\description{
Calls \link{plot.rivernetwork} and automatically zooms to a specified
  segment or vector of segments.  Not intended for any real mapping - just
  investigating and error checking.
}
\examples{
data(Kenai3)
plot(x=Kenai3)

# checking out a particularly messy region...
zoomtoseg(c(110,63), rivers=Kenai3)
}
\author{
Matt Tyers
}
